made adjustments so only mph needs adjusting

+++ RootDrivingChallenge/Driver.cs
@@ -11,20 +11,24 @@ namespace RootDrivingChallenge
         public double DriverMiles { get; set; }
         public string DriverName { get; set; }
         public double DriverTime { get; set; }
-        public double DriverSpeed { get; set; }
-        
+        public int DriverSpeed { get; set; }
+
+        //Calculate the total number of miles driven by the driver
         public double SumDriverTime(double tripDuration)
         {
             DriverTime += tripDuration;
             return DriverTime;
+            
         }
 
+        //Calculate the total number of miles driven by the driver
         public double SumDriverMiles(double tripMiles)
         {
             DriverMiles += tripMiles;
             return DriverMiles;
         }
 
+        //Calculate the MPH for the driver
         public int CalculateSpeed(double driverMiles, double driverTime)
         {
             int mph = Convert.ToInt32(Math.Round(driverMiles / DriverTime));
